Get all integers to 32 bits and correct resulting segfault
For consistency, I set all integers to uint32_t (except of course the ones
that should explicitly be on 8 bits). Following that, there was a segfault
caused by the blur functions. The problem was the following:

Declare uint32_t x = 0; uint32_t r = 10;
If you do x-r, the result is -10 BUT the result of (x-r) >= 0 is 1!
I understand that unsigned - unsigned is still unsigned, so the test fails.
It's just weird that the pure x-r operation (no test) gives -10. There's
probably implicit casts and priorities that I don't fully understand right now.
Anyway, I solved that by casting to signed integer before the test, which works
but might cause problems with other compilers.

+++ src/blur.c
@@ -11,6 +11,7 @@
  * #include
  *****************************************************************************/
 #include "blur.h"
+#include <stdio.h>
 
 /******************************************************************************
  * Private variables
@@ -19,10 +20,10 @@
 /******************************************************************************
  * Private functions
  *****************************************************************************/
-static void blur_x(uint16_t w, uint16_t h, float s[][h], float d[][h], 
-	uint16_t r);
-static void blur_y(uint16_t w, uint16_t h, float s[][h], float d[][h], 
-	uint16_t r);
+static void blur_x(uint32_t w, uint32_t h, float s[][h], float d[][h], 
+	uint32_t r);
+static void blur_y(uint32_t w, uint32_t h, float s[][h], float d[][h], 
+	uint32_t r);
 
 /**************************************************************************//**
  * Box blur main function.
@@ -33,7 +34,7 @@ static void blur_y(uint16_t w, uint16_t h, float s[][h], float d[][h],
  * @param d destination image (overwritten)
  * @param r blurring radius in pixels
  *****************************************************************************/
-void blur(uint16_t w, uint16_t h, float s[][h], float d[][h], uint16_t r)
+void blur(uint32_t w, uint32_t h, float s[][h], float d[][h], uint32_t r)
 {
 	float temp[w][h];
 	blur_x(w, h, s, temp, r);
@@ -49,10 +50,10 @@ void blur(uint16_t w, uint16_t h, float s[][h], float d[][h], uint16_t r)
  * @param d destination image (overwritten)
  * @param r blurring radius in pixels
  *****************************************************************************/
-void blur_x(uint16_t w, uint16_t h, float s[][h], float d[][h], uint16_t r)
+void blur_x(uint32_t w, uint32_t h, float s[][h], float d[][h], uint32_t r)
 {	
 	float sum;
-	uint16_t x, y, span;
+	uint32_t x, y, span;
 	
 	for (y=0; y<h; y++)
 	{
@@ -71,14 +72,19 @@ void blur_x(uint16_t w, uint16_t h, float s[][h], float d[][h], uint16_t r)
 		// picture. This wasn't really explained on blackpawn's page. And the
 		// span of the moving average isn't always equal to the radius, so it
 		// has to be computed as well.
+		// The casts to int32 are necessary, because subtracting two uint32
+		// always gives a positive result when doing the comparison. I think
+		// casting from unsigned to signed is compiler-dependent, though, so
+		// this might not be so safe.
 		for (x=1; x<w; x++)
 		{
-			sum -= (x - r - 1) >= 0 ? s[x-r-1][y] : 0;
-			sum += (x + r) < w ? s[x+r][y] : 0;
-            span = ((x + r) < w ? (x + r) : (w - 1))
-				 - ((x - r) >= 0 ? (x - r) : 0) + 1;
+			sum -= (int32_t)(x-r-1) >= 0 ? s[x-r-1][y] : 0;
+			sum += (x+r) < w ? s[x+r][y] : 0;
+            span = ((x+r) < w ? (x+r) : (w-1))
+				 - ((int32_t)(x-r) >= 0 ? (x-r) : 0) + 1;
             d[x][y] = sum / span;
 		}
+		
 	}
 }
 
@@ -91,34 +97,29 @@ void blur_x(uint16_t w, uint16_t h, float s[][h], float d[][h], uint16_t r)
  * @param d destination image (overwritten)
  * @param r blurring radius in pixels
  *****************************************************************************/
-void blur_y(uint16_t w, uint16_t h, float s[][h], float d[][h], uint16_t r)
+void blur_y(uint32_t w, uint32_t h, float s[][h], float d[][h], uint32_t r)
 {	
+	// See the comments in blur_x, it's exactly the same but working on the
+	// columns instead of the rows.
 	float sum;
-	uint16_t x, y, span;
+	uint32_t x, y, span;
 	
 	for (x=0; x<w; x++)
 	{
 		sum = 0;
 		
-		// In the blurred picture, the first pixel of each column is the 
-		// average of the source pixels between y=0 and y=r-1.
 		for (y=0; y<r; y++)
 		{
 			sum += s[x][y];
 		}
 		d[x][0] = sum / (r + 1);
 		
-		// The other pixels are computed with a moving average. Pixel values 
-		// are subtracted or added from the sum *only* if they are part of the
-		// picture. This wasn't really explained on blackpawn's page. And the
-		// span of the moving average isn't always equal to the radius, so it
-		// has to be computed as well.
-		for (y = 1; y < h; y++)
+		for (y=1; y<h; y++)
 		{
-			sum -= (y - r - 1) >= 0 ? s[x][y-r-1] : 0;
-			sum += (y + r) < h ? s[x][y+r] : 0;
-            span = ((y + r) < h ? (y + r) : (h - 1))
-				 - ((y - r) >= 0 ? (y - r) : 0) + 1;
+			sum -= (int32_t)(y-r-1) >= 0 ? s[x][y-r-1] : 0;
+			sum += (y+r) < h ? s[x][y+r] : 0;
+            span = ((y+r) < h ? (y+r) : (h-1))
+				 - ((int32_t)(y-r) >= 0 ? (y-r) : 0) + 1;
             d[x][y] = sum / span;
 		}
 	}
